Rename packages/runtime/runtime-definitions/src/garbageCollection.ts
