+++ packages/runtime/runtime-definitions/src/dataStoreContext.ts
@@ -26,7 +26,10 @@ import {
 import { IIdCompressor } from "@fluidframework/id-compressor";
 import { IProvideFluidDataStoreFactory } from "./dataStoreFactory";
 import { IProvideFluidDataStoreRegistry } from "./dataStoreRegistry";
-import { IGarbageCollectionData, IGarbageCollectionDetailsBase } from "./garbageCollection";
+import {
+	IGarbageCollectionData,
+	IGarbageCollectionDetailsBase,
+} from "./garbageCollectionDefinitions";
 import { IInboundSignalMessage } from "./protocol";
 import {
 	CreateChildSummarizerNodeParam,
