+++ packages/runtime/runtime-definitions/src/garbageCollectionDefinitions.ts
