+++ packages/runtime/runtime-definitions/src/index.ts
@@ -40,7 +40,7 @@ export {
 	gcTreeKey,
 	IGarbageCollectionData,
 	IGarbageCollectionDetailsBase,
-} from "./garbageCollection";
+} from "./garbageCollectionDefinitions";
 export {
 	IAttachMessage,
 	IEnvelope,
