+++ packages/runtime/runtime-definitions/src/summary.ts
@@ -11,7 +11,10 @@ import {
 	ISnapshotTree,
 	ITree,
 } from "@fluidframework/protocol-definitions";
-import { IGarbageCollectionData, IGarbageCollectionDetailsBase } from "./garbageCollection";
+import {
+	IGarbageCollectionData,
+	IGarbageCollectionDetailsBase,
+} from "./garbageCollectionDefinitions";
 
 /**
  * Contains the aggregation data from a Tree/Subtree.
